refactor(UpvoteService): Create models/database/vote package
