+++ lib/models/database/vote/upvote_state.dart
