+++ lib/services/database/post_upvote_service.dart
@@ -3,8 +3,8 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/post/post_data.dart";
 import "package:proxima/models/database/post/post_firestore.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
-import "package:proxima/models/database/upvote_state.dart";
 import "package:proxima/models/database/user/user_id_firestore.dart";
+import "package:proxima/models/database/vote/upvote_state.dart";
 import "package:proxima/services/database/firestore_service.dart";
 
 /// This repository service is responsible for handling the upvotes of posts
